faiss_index.py
  
import faiss  
import numpy as np  

def search_documents(query, text):  
    index = faiss.IndexFlatL2(768)  # Placeholder, needs real embeddings  
    query_vector = np.random.rand(1, 768).astype("float32")  
    index.add(np.random.rand(10, 768).astype("float32"))  
    _, I = index.search(query_vector, 1)  
    return f"Best matching result: {I[0][0]}"  
    