requirements.txt
  
streamlit  
langchain  
faiss-cpu  
openai  
pytesseract  
pdfplumber  
Pillow  
    